Rename status_row to status_message.rs and make unsaved change message orange as a warning not red as an error.
