+++ src/views/mod.rs
@@ -5,6 +5,6 @@ pub mod info_row;
 pub mod layout_selector;
 pub mod main_row;
 pub mod pin_state;
-pub mod status_row;
+pub mod status_message;
 pub mod version;
 pub mod waveform;
